requirements.txt
Python 3.7+

NumPy

SciPy

Matplotlib

PyTorch

Jupyter Notebook